Refactor phase 3d: split the admin API routes into a package
webapp/admin/api.py (2,105 lines) becomes the webapp/admin/api/ package:
13 modules plus a 95-line __init__, all within the 400-line guidance.

Unlike 3b and 3c, this file was 21 ordinary top-level definitions
decorated with @api_bp.route rather than handlers nested inside one
giant register(), so every definition moved verbatim.

  blueprint, county, hostinfo, motion       leaves
  display_data           -> motion
  routes_geometry        -> blueprint, county
  routes_alert_detail    -> blueprint, county, display_data
  routes_alert_export    -> blueprint, display_data
  routes_alerts_list     -> blueprint, county
  routes_boundaries      -> blueprint
  routes_system          -> blueprint, hostinfo
  routes_system_history  -> blueprint
  routes_smart           -> blueprint

register_api_routes(app, logger) is unchanged and is the only name the
package re-exports.

Import blocks are derived rather than hand-written, which cost 127 ruff
errors to learn. The free-variable set comes from symtable, not from
counting ast.Name nodes: a local variable named `desc` inside
_extract_alert_display_data reads as a use of `from sqlalchemy import
desc` under name counting. Three genuinely dead imports fell out as a
result.

blueprint.py passes __package__ rather than __name__ so the blueprint
keeps its import_name. root_path does move, since the name now resolves
to a package; a test pins that it is inert (no template or static
folder).

Two test files read the API source as text and would have broken on the
move — worse, pointed at a shim they would have passed vacuously. Both
now scan the package directory. tests/test_api_package.py adds the
structural guards and exercises the real _detect_county_wide against a
stub alert instead of the local mirror the old tests reimplemented; both
heuristics are mutation-checked.

Verification: 21/21 definitions ast.dump()-identical, every non-blank
line placed exactly once, URL map unchanged at 549 rules / 0 differences
against a worktree at the pre-split commit, full suite 2256 passed.
