+++ webapp/admin/api/routes_alert_export.py
@@ -1,0 +1,265 @@
+"""
+EAS Station - Emergency Alert System
+Copyright (c) 2025-2026 EAS Station, LLC (KR8MER)
+
+This file is part of EAS Station.
+
+EAS Station is dual-licensed software:
+- GNU Affero General Public License v3 (AGPL-3.0) for open-source use
+- Commercial License for proprietary use
+
+You should have received a copy of both licenses with this software.
+For more information, see LICENSE and LICENSE-COMMERCIAL files.
+
+IMPORTANT: This software cannot be rebranded or have attribution removed.
+See NOTICE file for complete terms.
+
+Repository: https://github.com/KR8MER/eas-station
+"""
+
+from __future__ import annotations
+
+"""Exporting one alert: PDF, social-share image, and IPAWS audio.
+
+The three endpoints that hand an alert to the user as a file rather than a
+page. Each re-reads the alert itself rather than sharing state with the detail
+page, so a direct link to an export works without loading the page first.
+"""
+
+import os
+from datetime import datetime
+
+from flask import flash, redirect, request, url_for, Response
+
+from app_core.extensions import db
+from app_core.models import Boundary, CAPAlert, Intersection
+from app_core.eas_storage import format_local_datetime
+from app_utils.pdf_generator import generate_pdf_document
+
+from ..coverage import calculate_coverage_percentages
+
+from .blueprint import api_bp
+from .display_data import _extract_alert_display_data
+
+
+@api_bp.route('/alerts/<int:alert_id>/export.pdf')
+def alert_detail_pdf(alert_id):
+    """Generate archival PDF for a specific alert - server-side from database."""
+    try:
+        alert = CAPAlert.query.get_or_404(alert_id)
+
+        # Build PDF sections
+        sections = []
+
+        # Alert Information Section
+        alert_info = [
+            f"Event: {alert.event or 'N/A'}",
+            f"Severity: {alert.severity or 'N/A'}",
+            f"Status: {alert.status or 'N/A'}",
+            f"Message Type: {alert.message_type or 'N/A'}",
+            f"Urgency: {alert.urgency or 'N/A'}",
+            f"Certainty: {alert.certainty or 'N/A'}",
+            f"Identifier: {alert.identifier or 'N/A'}",
+        ]
+
+        if alert.sent:
+            alert_info.append(f"Sent: {format_local_datetime(alert.sent, include_utc=True)}")
+        if alert.expires:
+            alert_info.append(f"Expires: {format_local_datetime(alert.expires, include_utc=True)}")
+
+        sections.append({
+            'heading': 'Alert Information',
+            'content': alert_info,
+        })
+
+        # Description Section
+        if alert.headline:
+            sections.append({
+                'heading': 'Headline',
+                'content': alert.headline,
+            })
+
+        if alert.description:
+            sections.append({
+                'heading': 'Description',
+                'content': alert.description,
+            })
+
+        if alert.instruction:
+            sections.append({
+                'heading': 'Instructions',
+                'content': alert.instruction,
+            })
+
+        # Area Information — geocode fields were removed from the model;
+        # extract them from the stored raw CAP JSON payload instead.
+        _props = (alert.raw_json or {}).get('properties', {}) if alert.raw_json else {}
+        _geocode = _props.get('geocode', {}) or {}
+        geocode_same = _geocode.get('SAME', [])
+        geocode_ugc = _geocode.get('UGC', [])
+
+        area_info = []
+        if alert.area_desc:
+            area_info.append(f"Area Description: {alert.area_desc}")
+        if geocode_same:
+            area_info.append(f"SAME Codes: {', '.join(geocode_same)}")
+        if geocode_ugc:
+            area_info.append(f"UGC Codes: {', '.join(geocode_ugc)}")
+
+        if area_info:
+            sections.append({
+                'heading': 'Affected Areas',
+                'content': area_info,
+            })
+
+        # Source Information — sender fields were removed from the model;
+        # extract them from raw_json instead.
+        sender_name = _props.get('senderName', '') or ''
+        sender_id = _props.get('sender', '') or ''
+
+        source_info = []
+        if sender_name:
+            source_info.append(f"Sender: {sender_name}")
+        if sender_id:
+            source_info.append(f"Sender ID: {sender_id}")
+        if alert.source:
+            source_info.append(f"Source: {alert.source}")
+
+        if source_info:
+            sections.append({
+                'heading': 'Source Information',
+                'content': source_info,
+            })
+
+        # Generate PDF
+        pdf_bytes = generate_pdf_document(
+            title=f"Alert Detail Report - {alert.event or 'Alert'}",
+            sections=sections,
+            subtitle=f"Alert ID: {alert_id}",
+            footer_text="Generated by EAS Station™ — Emergency Alert System Platform"
+        )
+
+        # Return as downloadable PDF
+        response = Response(pdf_bytes, mimetype="application/pdf")
+        response.headers["Content-Disposition"] = (
+            f"inline; filename=alert_{alert_id}_{datetime.now().strftime('%Y%m%d')}.pdf"
+        )
+        return response
+
+    except Exception as exc:
+        api_bp.logger.error('Error generating alert PDF: %s', exc, exc_info=True)
+        flash('Error generating PDF. Please try again.', 'error')
+        return redirect(url_for('api.alert_detail', alert_id=alert_id))
+
+_SOCIAL_IMAGE_RATIOS  = {'landscape', 'square', 'portrait', 'story'}
+
+_SOCIAL_IMAGE_FORMATS = {'png', 'webp'}
+
+@api_bp.route('/alerts/<int:alert_id>/export-image.png')
+def alert_detail_image(alert_id):
+    """Generate a social-share image for an alert.
+
+    Accepts three optional query arguments:
+
+    * ``ratio`` — one of ``landscape`` (1200×630, FB/X/LinkedIn
+      open-graph cards), ``square`` (1080×1080 — Instagram, Mastodon),
+      ``portrait`` (1080×1350 — Instagram 4:5), or ``story``
+      (1080×1920 — IG / TikTok / Snap Stories & Reels).
+    * ``format`` — ``png`` (default, universal) or ``webp`` (~30%
+      smaller at equivalent quality; supported by every major social
+      platform).
+    * ``scale`` — output upscale factor, 1–3 (default ``2``).  Social
+      platforms re-encode uploads to JPEG and display them downscaled;
+      a 2× upload keeps that compression from reading as grain in the
+      feed.  Pass ``scale=1`` for the native canvas size.
+
+    Unknown values quietly fall back to ``landscape`` / ``png`` / ``2``.
+    The file name and Content-Type are set from the chosen format so the
+    browser saves the file with the right extension.
+    """
+    try:
+        from app_utils.image_export import generate_alert_image
+        from app_core.location import get_location_settings
+
+        alert = CAPAlert.query.get_or_404(alert_id)
+
+        ratio = (request.args.get('ratio') or 'landscape').strip().lower()
+        if ratio not in _SOCIAL_IMAGE_RATIOS:
+            ratio = 'landscape'
+
+        fmt = (request.args.get('format') or 'png').strip().lower()
+        if fmt not in _SOCIAL_IMAGE_FORMATS:
+            fmt = 'png'
+
+        try:
+            scale = float(request.args.get('scale', 2))
+        except (TypeError, ValueError):
+            scale = 2.0
+        scale = max(1.0, min(scale, 3.0))
+
+        intersections = db.session.query(Intersection, Boundary).join(
+            Boundary, Intersection.boundary_id == Boundary.id
+        ).filter(Intersection.cap_alert_id == alert_id).all()
+
+        coverage_data = calculate_coverage_percentages(alert_id, intersections)
+        ipaws_data    = _extract_alert_display_data(alert)
+
+        try:
+            location_settings = get_location_settings()
+        except Exception:
+            location_settings = {}
+
+        image_bytes = generate_alert_image(
+            alert, coverage_data, ipaws_data, location_settings,
+            aspect_ratio=ratio, image_format=fmt, scale=scale,
+        )
+
+        mimetype = 'image/webp' if fmt == 'webp' else 'image/png'
+        response = Response(image_bytes, mimetype=mimetype)
+        safe_event = (alert.event or 'alert').replace(' ', '_').lower()
+        # Include the ratio + format in the filename so saving multiple
+        # variants doesn't collide on disk.
+        response.headers['Content-Disposition'] = (
+            f'attachment; filename=alert_{alert_id}_{safe_event}_{ratio}.{fmt}'
+        )
+        response.headers['Cache-Control'] = 'no-cache'
+        return response
+
+    except Exception as exc:
+        api_bp.logger.error('Error generating alert image: %s', exc, exc_info=True)
+        flash('Error generating image. Please try again.', 'error')
+        return redirect(url_for('api.alert_detail', alert_id=alert_id))
+
+@api_bp.route('/alerts/<int:alert_id>/ipaws_audio')
+def ipaws_original_audio(alert_id):
+    """Serve the original IPAWS audio file for an alert."""
+    from flask import send_file
+    alert = CAPAlert.query.get_or_404(alert_id)
+    filename = getattr(alert, 'ipaws_audio_url', None)
+    if not filename:
+        return Response('No IPAWS audio available for this alert', status=404)
+
+    # Prevent path traversal by ensuring filename contains no path separators
+    safe_filename = os.path.basename(filename)
+    if not safe_filename or safe_filename != filename:
+        return Response('Invalid audio filename', status=400)
+
+    eas_output = os.getenv('EAS_OUTPUT_DIR') or os.path.join(
+        os.getenv('EAS_STATIC_DIR', os.path.join(os.getcwd(), 'static')),
+        'eas_messages',
+    )
+    filepath = os.path.join(eas_output, safe_filename)
+    
+    # Verify the resolved path is within the output directory
+    try:
+        real_filepath = os.path.realpath(filepath)
+        real_output = os.path.realpath(eas_output)
+        if not real_filepath.startswith(real_output + os.sep):
+            return Response('Invalid audio file path', status=400)
+    except (OSError, ValueError):
+        return Response('Invalid audio file path', status=400)
+    
+    if not os.path.isfile(real_filepath):
+        return Response('Audio file not found on disk', status=404)
+
+    return send_file(real_filepath, mimetype='audio/mpeg', download_name=safe_filename)
